Commonize details for instruction `Operand` downwards
