+++ src/lib.rs
@@ -103,12 +103,13 @@
 
 #![no_std]
 
-#[cfg(feature="use-serde")]
-#[macro_use] extern crate serde_derive;
-#[cfg(feature="use-serde")]
+#[cfg(feature = "use-serde")]
+#[macro_use]
+extern crate serde_derive;
+#[cfg(feature = "use-serde")]
 extern crate serde;
 
-#[cfg(feature="std")]
+#[cfg(feature = "std")]
 extern crate alloc;
 
 pub mod long_mode;
@@ -120,19 +121,19 @@ pub use protected_mode::Arch as x86_32;
 
 pub mod real_mode;
 pub use real_mode::Arch as x86_16;
+use yaxpeax_arch::Instruction;
+
+use crate::safer_unchecked::unreachable_kinda_unchecked as unreachable_unchecked;
 
 mod display;
 mod safer_unchecked;
 
 const MEM_SIZE_STRINGS: [&'static str; 64] = [
-    "byte", "word", "BUG", "dword", "ptr", "far", "BUG", "qword",
-    "BUG", "mword", "BUG", "BUG", "BUG", "BUG", "BUG", "xmmword",
-    "BUG", "BUG", "BUG", "BUG", "BUG", "BUG", "BUG", "BUG",
-    "BUG", "BUG", "BUG", "BUG", "BUG", "BUG", "BUG", "ymmword",
-    "BUG", "BUG", "BUG", "BUG", "BUG", "BUG", "BUG", "BUG",
-    "BUG", "BUG", "BUG", "BUG", "BUG", "BUG", "BUG", "BUG",
-    "BUG", "BUG", "BUG", "BUG", "BUG", "BUG", "BUG", "BUG",
-    "BUG", "BUG", "BUG", "BUG", "BUG", "BUG", "ptr", "zmmword",
+    "byte", "word", "BUG", "dword", "ptr", "far", "BUG", "qword", "BUG", "mword", "BUG", "BUG",
+    "BUG", "BUG", "BUG", "xmmword", "BUG", "BUG", "BUG", "BUG", "BUG", "BUG", "BUG", "BUG", "BUG",
+    "BUG", "BUG", "BUG", "BUG", "BUG", "BUG", "ymmword", "BUG", "BUG", "BUG", "BUG", "BUG", "BUG",
+    "BUG", "BUG", "BUG", "BUG", "BUG", "BUG", "BUG", "BUG", "BUG", "BUG", "BUG", "BUG", "BUG",
+    "BUG", "BUG", "BUG", "BUG", "BUG", "BUG", "BUG", "BUG", "BUG", "BUG", "BUG", "ptr", "zmmword",
 ];
 
 pub struct MemoryAccessSize {
@@ -193,6 +194,750 @@ impl core::fmt::Display for MemoryAccessSize {
 impl core::fmt::Debug for MemoryAccessSize {
     fn fmt(&self, f: &mut core::fmt::Formatter) -> core::fmt::Result {
         core::fmt::Display::fmt(self, f)
+    }
+}
+
+pub trait X86Instruction: Instruction {
+    /// Get the opcode of this instruction
+    fn opcode(&self) -> Opcode;
+
+    /// get the `Operand` at the provided index.
+    ///
+    /// panics if the index is `>= 4`.
+    fn operand(&self, i: u8) -> Operand;
+
+    /// get the number of operands in this instruction. useful in iterating an instruction's
+    /// operands generically.
+    fn operand_count(&self) -> u8;
+
+    /// get the `Segment` that will *actually* be used for accessing the operand at index `i`.
+    ///
+    /// `stos`, `lods`, `movs`, and `cmps` specifically name some segments for use regardless of
+    /// prefixes.
+    fn segment_override_for_op(&self, op: u8) -> Option<Segment>;
+}
+
+impl X86Instruction for crate::long_mode::Instruction {
+    fn opcode(&self) -> Opcode {
+        self.opcode()
+    }
+
+    fn operand(&self, i: u8) -> Operand {
+        self.operand(i)
+    }
+
+    fn operand_count(&self) -> u8 {
+        self.operand_count()
+    }
+
+    fn segment_override_for_op(&self, op: u8) -> Option<Segment> {
+        self.segment_override_for_op(op)
+    }
+}
+
+impl X86Instruction for crate::protected_mode::Instruction {
+    fn opcode(&self) -> Opcode {
+        self.opcode()
+    }
+
+    fn operand(&self, i: u8) -> Operand {
+        self.operand(i)
+    }
+
+    fn operand_count(&self) -> u8 {
+        self.operand_count()
+    }
+
+    fn segment_override_for_op(&self, op: u8) -> Option<Segment> {
+        self.segment_override_for_op(op)
+    }
+}
+
+impl X86Instruction for crate::real_mode::Instruction {
+    fn opcode(&self) -> Opcode {
+        self.opcode()
+    }
+
+    fn operand(&self, i: u8) -> Operand {
+        self.operand(i)
+    }
+
+    fn operand_count(&self) -> u8 {
+        self.operand_count()
+    }
+
+    fn segment_override_for_op(&self, op: u8) -> Option<Segment> {
+        self.segment_override_for_op(op)
+    }
+}
+
+const REGISTER_CLASS_NAMES: &[&'static str] = &[
+    "qword",
+    "BUG. PLEASE REPORT.",
+    "dword",
+    "BUG. PLEASE REPORT.",
+    "word",
+    "BUG. PLEASE REPORT.",
+    "byte",
+    "BUG. PLEASE REPORT.",
+    "rex-byte",
+    "BUG. PLEASE REPORT.",
+    "cr",
+    "BUG. PLEASE REPORT.",
+    "dr",
+    "BUG. PLEASE REPORT.",
+    "segment",
+    "xmm",
+    "BUG. PLEASE REPORT.",
+    "BUG. PLEASE REPORT.",
+    "BUG. PLEASE REPORT.",
+    "ymm",
+    "BUG. PLEASE REPORT.",
+    "BUG. PLEASE REPORT.",
+    "BUG. PLEASE REPORT.",
+    "zmm",
+    "BUG. PLEASE REPORT.",
+    "BUG. PLEASE REPORT.",
+    "BUG. PLEASE REPORT.",
+    "x87-stack",
+    "mmx",
+    "k",
+    "eip",
+    "rip",
+    "eflags",
+    "rflags",
+];
+
+/// an `x86_64` register class - `qword`, `dword`, `xmmword`, `segment`, and so on.
+///
+/// this is mostly useful for comparing a `RegSpec`'s [`RegSpec::class()`] with a constant out of
+/// [`register_class`].
+#[cfg_attr(feature = "use-serde", derive(Serialize, Deserialize))]
+#[derive(Copy, Clone, Debug, Ord, PartialOrd, Eq, PartialEq, Hash)]
+pub struct RegisterClass {
+    kind: RegisterBank,
+}
+
+impl RegisterClass {
+    /// return a human-friendly name for this register class
+    pub fn name(&self) -> &'static str {
+        REGISTER_CLASS_NAMES[self.kind as usize]
+    }
+
+    /// return the size of this register class, in bytes
+    pub fn width(&self) -> u8 {
+        match self.kind {
+            RegisterBank::Q => 8,
+            RegisterBank::D => 4,
+            RegisterBank::W => 2,
+            RegisterBank::B | RegisterBank::rB => 1,
+            RegisterBank::CR | RegisterBank::DR => 8,
+            RegisterBank::S => 2,
+            RegisterBank::EIP => 4,
+            RegisterBank::RIP => 8,
+            RegisterBank::EFlags => 4,
+            RegisterBank::RFlags => 8,
+            RegisterBank::X => 16,
+            RegisterBank::Y => 32,
+            RegisterBank::Z => 64,
+            RegisterBank::ST => 10,
+            RegisterBank::MM => 8,
+            RegisterBank::K => 8,
+        }
+    }
+}
+
+#[allow(non_camel_case_types)]
+#[cfg_attr(feature = "use-serde", derive(Serialize, Deserialize))]
+#[derive(Copy, Clone, Debug, Ord, PartialOrd, Eq, PartialEq, Hash)]
+#[repr(u8)]
+enum RegisterBank {
+    Q = 0,
+    D = 2,
+    W = 4,
+    B = 6,
+    rB = 8, // Quadword, Dword, Word, Byte
+    CR = 10,
+    DR = 12,
+    S = 14,
+    EIP = 30,
+    RIP = 31,
+    EFlags = 32,
+    RFlags = 33, // Control reg, Debug reg, Selector, ...
+    X = 15,
+    Y = 19,
+    Z = 23, // XMM, YMM, ZMM
+    ST = 27,
+    MM = 28, // ST, MM regs (x87, mmx)
+    K = 29,  // AVX512 mask registers
+}
+
+/// the segment register used by the corresponding instruction.
+///
+/// typically this will be `ds` but can be overridden. some instructions have specific segment
+/// registers used regardless of segment prefixes, and in these cases `yaxpeax-x86` will report the
+/// actual segment register a physical processor would use.
+#[derive(Copy, Clone, Debug, Eq, PartialEq, Hash)]
+pub enum Segment {
+    DS = 0,
+    CS,
+    ES,
+    FS,
+    GS,
+    SS,
+}
+
+/// an `x86` register, including its number and type. if `fmt` is enabled, name too.
+///
+/// ```
+/// use yaxpeax_x86::{RegSpec, long_mode::register_class};
+///
+/// assert_eq!(RegSpec::ecx().num(), 1);
+/// assert_eq!(RegSpec::ecx().class(), register_class::D);
+/// ```
+///
+/// some registers have classes of their own, and only one member: `eip` and `eflags`.
+#[cfg_attr(feature = "use-serde", derive(Serialize, Deserialize))]
+#[derive(Copy, Clone, Debug, PartialOrd, Ord, Eq, PartialEq)]
+pub struct RegSpec {
+    num: u8,
+    bank: RegisterBank,
+}
+
+use core::hash::Hash;
+use core::hash::Hasher;
+impl Hash for RegSpec {
+    fn hash<H: Hasher>(&self, state: &mut H) {
+        let code = ((self.bank as u16) << 8) | (self.num as u16);
+        code.hash(state);
+    }
+}
+
+macro_rules! register {
+    ($bank:ident, $name:ident => $num:expr, $($tail:tt)+) => {
+        #[inline]
+        pub const fn $name() -> RegSpec {
+            RegSpec { bank: RegisterBank::$bank, num: $num }
+        }
+
+        register!($bank, $($tail)*);
+    };
+    ($bank:ident, $name:ident => $num:expr) => {
+        #[inline]
+        pub const fn $name() -> RegSpec {
+            RegSpec { bank: RegisterBank::$bank, num: $num }
+        }
+    };
+}
+
+#[inline]
+fn width_to_gp_reg_bank(width: u8, rex: bool) -> RegisterBank {
+    match width {
+        1 => {
+            return if rex {
+                RegisterBank::rB
+            } else {
+                RegisterBank::B
+            }
+        }
+        2 => return RegisterBank::W,
+        4 => return RegisterBank::D,
+        8 => return RegisterBank::Q,
+        _ => unsafe {
+            unreachable_unchecked();
+        },
+    }
+}
+
+#[allow(non_snake_case)]
+impl RegSpec {
+    /// the register `rip`. this register is in the class `rip`, which contains only it.
+    pub const RIP: RegSpec = RegSpec::rip();
+    /// the register `eip`. this register is in the class `rip`, which contains only it.
+    pub const EIP: RegSpec = RegSpec::eip();
+
+    /// the number of this register in its `RegisterClass`.
+    ///
+    /// for many registers this is a number in the name, but for registers harkening back to
+    /// `x86_32`, the first eight registers are `rax`, `rcx`, `rdx`, `rbx`, `rsp`, `rbp`, `rsi`,
+    /// and `rdi` (or `eXX` for the 32-bit forms, `XX` for 16-bit forms).
+    pub fn num(&self) -> u8 {
+        self.num
+    }
+
+    /// the class of register this register is in.
+    ///
+    /// this corresponds to the register's size, but is by the register's usage in the instruction
+    /// set; `rax` and `mm0` are the same size, but different classes (`Q`(word) and `MM` (mmx)
+    /// respectively).
+    pub fn class(&self) -> RegisterClass {
+        RegisterClass { kind: self.bank }
+    }
+
+    #[cfg(feature = "fmt")]
+    /// return a human-friendly name for this register. the returned name is the same as would be
+    /// used to render this register in an instruction.
+    pub fn name(&self) -> &'static str {
+        display::regspec_label(self)
+    }
+
+    /// construct a `RegSpec` for x87 register `st(num)`
+    #[inline]
+    pub fn st(num: u8) -> RegSpec {
+        if num >= 8 {
+            panic!("invalid x87 reg st({})", num);
+        }
+
+        RegSpec {
+            num,
+            bank: RegisterBank::ST,
+        }
+    }
+
+    /// construct a `RegSpec` for xmm reg `num`
+    #[inline]
+    pub fn xmm(num: u8) -> RegSpec {
+        if num >= 32 {
+            panic!("invalid x86 xmm reg {}", num);
+        }
+
+        RegSpec {
+            num,
+            bank: RegisterBank::X,
+        }
+    }
+
+    /// construct a `RegSpec` for ymm reg `num`
+    #[inline]
+    pub fn ymm(num: u8) -> RegSpec {
+        if num >= 32 {
+            panic!("invalid x86 ymm reg {}", num);
+        }
+
+        RegSpec {
+            num,
+            bank: RegisterBank::Y,
+        }
+    }
+
+    /// construct a `RegSpec` for zmm reg `num`
+    #[inline]
+    pub fn zmm(num: u8) -> RegSpec {
+        if num >= 32 {
+            panic!("invalid x86 zmm reg {}", num);
+        }
+
+        RegSpec {
+            num,
+            bank: RegisterBank::Z,
+        }
+    }
+
+    /// construct a `RegSpec` for qword reg `num`
+    #[inline]
+    pub fn q(num: u8) -> RegSpec {
+        if num >= 16 {
+            panic!("invalid x86 qword reg {}", num);
+        }
+
+        RegSpec {
+            num,
+            bank: RegisterBank::Q,
+        }
+    }
+
+    /// construct a `RegSpec` for mask reg `num`
+    #[inline]
+    pub fn mask(num: u8) -> RegSpec {
+        if num >= 8 {
+            panic!("invalid x86 mask reg {}", num);
+        }
+
+        RegSpec {
+            num,
+            bank: RegisterBank::K,
+        }
+    }
+
+    /// construct a `RegSpec` for dword reg `num`
+    #[inline]
+    pub fn d(num: u8) -> RegSpec {
+        if num >= 16 {
+            panic!("invalid x86 dword reg {}", num);
+        }
+
+        RegSpec {
+            num,
+            bank: RegisterBank::D,
+        }
+    }
+
+    /// construct a `RegSpec` for word reg `num`
+    #[inline]
+    pub fn w(num: u8) -> RegSpec {
+        if num >= 16 {
+            panic!("invalid x86 word reg {}", num);
+        }
+
+        RegSpec {
+            num,
+            bank: RegisterBank::W,
+        }
+    }
+
+    /// construct a `RegSpec` for non-rex byte reg `num`
+    #[inline]
+    pub fn rb(num: u8) -> RegSpec {
+        if num >= 16 {
+            panic!("invalid x86 rex-byte reg {}", num);
+        }
+
+        RegSpec {
+            num,
+            bank: RegisterBank::rB,
+        }
+    }
+
+    /// construct a `RegSpec` for non-rex byte reg `num`
+    #[inline]
+    pub fn b(num: u8) -> RegSpec {
+        if num >= 8 {
+            panic!("invalid x86 non-rex byte reg {}", num);
+        }
+
+        RegSpec {
+            num,
+            bank: RegisterBank::B,
+        }
+    }
+
+    #[inline]
+    fn from_parts(num: u8, extended: bool, bank: RegisterBank) -> RegSpec {
+        RegSpec {
+            num: num + if extended { 0b1000 } else { 0 },
+            bank: bank,
+        }
+    }
+
+    #[inline]
+    fn gp_from_parts(num: u8, extended: bool, width: u8, rex: bool) -> RegSpec {
+        RegSpec {
+            num: num + if extended { 0b1000 } else { 0 },
+            bank: width_to_gp_reg_bank(width, rex),
+        }
+    }
+
+    register!(RIP, rip => 0);
+    register!(EIP, eip => 0);
+
+    register!(RFlags, rflags => 0);
+    register!(EFlags, eflags => 0);
+
+    register!(S, es => 0, cs => 1, ss => 2, ds => 3, fs => 4, gs => 5);
+
+    register!(Q,
+        rax => 0, rcx => 1, rdx => 2, rbx => 3,
+        rsp => 4, rbp => 5, rsi => 6, rdi => 7,
+        r8 => 8, r9 => 9, r10 => 10, r11 => 11,
+        r12 => 8, r13 => 9, r14 => 14, r15 => 15
+    );
+
+    register!(D,
+        eax => 0, ecx => 1, edx => 2, ebx => 3,
+        esp => 4, ebp => 5, esi => 6, edi => 7,
+        r8d => 8, r9d => 9, r10d => 10, r11d => 11,
+        r12d => 8, r13d => 9, r14d => 14, r15d => 15
+    );
+
+    register!(W,
+        ax => 0, cx => 1, dx => 2, bx => 3,
+        sp => 4, bp => 5, si => 6, di => 7,
+        r8w => 8, r9w => 9, r10w => 10, r11w => 11,
+        r12w => 8, r13w => 9, r14w => 14, r15w => 15
+    );
+
+    register!(B,
+        al => 0, cl => 1, dl => 2, bl => 3,
+        ah => 4, ch => 5, dh => 6, bh => 7
+    );
+
+    register!(rB,
+        spl => 4, bpl => 5, sil => 6, dil => 7,
+        r8b => 8, r9b => 9, r10b => 10, r11b => 11,
+        r12b => 8, r13b => 9, r14b => 14, r15b => 15
+    );
+
+    #[inline]
+    pub const fn zmm0() -> RegSpec {
+        RegSpec {
+            bank: RegisterBank::Z,
+            num: 0,
+        }
+    }
+
+    #[inline]
+    pub const fn ymm0() -> RegSpec {
+        RegSpec {
+            bank: RegisterBank::Y,
+            num: 0,
+        }
+    }
+
+    #[inline]
+    pub const fn xmm0() -> RegSpec {
+        RegSpec {
+            bank: RegisterBank::X,
+            num: 0,
+        }
+    }
+
+    #[inline]
+    pub const fn st0() -> RegSpec {
+        RegSpec {
+            bank: RegisterBank::ST,
+            num: 0,
+        }
+    }
+
+    #[inline]
+    pub const fn mm0() -> RegSpec {
+        RegSpec {
+            bank: RegisterBank::MM,
+            num: 0,
+        }
+    }
+
+    /// return the size of this register, in bytes.
+    #[inline]
+    pub fn width(&self) -> u8 {
+        self.class().width()
+    }
+}
+
+/// an `avx512` merging mode.
+///
+/// the behavior for non-`avx512` instructions is equivalent to `merge`.  `zero` is only useful in
+/// conjunction with a mask register, where bits specified in the mask register correspond to
+/// unmodified items in the instruction's desination.
+#[derive(Debug, Copy, Clone, PartialEq, Eq)]
+pub enum MergeMode {
+    Merge,
+    Zero,
+}
+impl From<bool> for MergeMode {
+    fn from(b: bool) -> Self {
+        if b {
+            MergeMode::Zero
+        } else {
+            MergeMode::Merge
+        }
+    }
+}
+/// an `avx512` custom rounding mode.
+#[derive(Debug, Copy, Clone, PartialEq, Eq)]
+pub enum SaeMode {
+    RoundNearest,
+    RoundDown,
+    RoundUp,
+    RoundZero,
+}
+const SAE_MODES: [SaeMode; 4] = [
+    SaeMode::RoundNearest,
+    SaeMode::RoundDown,
+    SaeMode::RoundUp,
+    SaeMode::RoundZero,
+];
+impl SaeMode {
+    /// a human-friendly label for this `SaeMode`:
+    ///
+    /// ```
+    /// use yaxpeax_x86::long_mode::SaeMode;
+    ///
+    /// assert_eq!(SaeMode::RoundNearest.label(), "{rne-sae}");
+    /// assert_eq!(SaeMode::RoundDown.label(), "{rd-sae}");
+    /// assert_eq!(SaeMode::RoundUp.label(), "{ru-sae}");
+    /// assert_eq!(SaeMode::RoundZero.label(), "{rz-sae}");
+    /// ```
+    pub fn label(&self) -> &'static str {
+        match self {
+            SaeMode::RoundNearest => "{rne-sae}",
+            SaeMode::RoundDown => "{rd-sae}",
+            SaeMode::RoundUp => "{ru-sae}",
+            SaeMode::RoundZero => "{rz-sae}",
+        }
+    }
+
+    fn from(l: bool, lp: bool) -> Self {
+        let mut idx = 0;
+        if l {
+            idx |= 1;
+        }
+        if lp {
+            idx |= 2;
+        }
+        SAE_MODES[idx]
+    }
+}
+
+/// an operand for an `x86` instruction.
+///
+/// `Operand::Nothing` should be unreachable in practice; any such instructions should have an
+/// operand count of 0 (or at least one fewer than the `Nothing` operand's position).
+#[derive(Clone, Debug, PartialEq)]
+#[non_exhaustive]
+pub enum Operand {
+    /// a sign-extended byte
+    ImmediateI8(i8),
+    /// a zero-extended byte
+    ImmediateU8(u8),
+    /// a sign-extended word
+    ImmediateI16(i16),
+    /// a zero-extended word
+    ImmediateU16(u16),
+    /// a sign-extended dword
+    ImmediateI32(i32),
+    /// a zero-extended dword
+    ImmediateU32(u32),
+    /// a sign-extended qword
+    ImmediateI64(i64),
+    /// a zero-extended qword
+    ImmediateU64(u64),
+    /// a bare register operand, such as `rcx`.
+    Register(RegSpec),
+    /// an `avx512` register operand with optional mask register and merge mode, such as
+    /// `zmm3{k4}{z}`.
+    ///
+    /// if the mask register is `k0`, there is no masking applied, and the default x86 operation is
+    /// `MergeMode::Merge`.
+    RegisterMaskMerge(RegSpec, RegSpec, MergeMode),
+    /// an `avx512` register operand with optional mask register, merge mode, and suppressed
+    /// exceptions, such as `zmm3{k4}{z}{rd-sae}`.
+    ///
+    /// if the mask register is `k0`, there is no masking applied, and the default x86 operation is
+    /// `MergeMode::Merge`.
+    RegisterMaskMergeSae(RegSpec, RegSpec, MergeMode, SaeMode),
+    /// an `avx512` register operand with optional mask register, merge mode, and suppressed
+    /// exceptions, with no overridden rounding mode, such as `zmm3{k4}{z}{sae}`.
+    ///
+    /// if the mask register is `k0`, there is no masking applied, and the default x86 operation is
+    /// `MergeMode::Merge`.
+    RegisterMaskMergeSaeNoround(RegSpec, RegSpec, MergeMode),
+    /// a memory access to a literal word address. it's extremely rare that a well-formed x86
+    /// instruction uses this mode. as an example, `[0x1133]`
+    DisplacementU16(u16),
+    /// a memory access to a literal qword address. it's relatively rare that a well-formed x86
+    /// instruction uses this mode, but plausibe. for example, `fs:[0x14]`. segment overrides,
+    /// however, are maintained on the instruction itself.
+    DisplacementU32(u32),
+    /// a memory access to a literal qword address. it's relatively rare that a well-formed x86
+    /// instruction uses this mode, but plausibe. for example, `gs:[0x14]`. segment overrides,
+    /// however, are maintained on the instruction itself.
+    DisplacementU64(u64),
+    /// a simple dereference of the address held in some register. for example: `[esi]`.
+    RegDeref(RegSpec),
+    /// a dereference of the address held in some register with offset. for example: `[esi + 0x14]`.
+    RegDisp(RegSpec, i32),
+    /// a dereference of the address held in some register scaled by 1, 2, 4, or 8. this is almost always used with the `lea` instruction. for example: `[edx * 4]`.
+    RegScale(RegSpec, u8),
+    /// a dereference of the address from summing two registers. for example: `[ebp + rax]`
+    RegIndexBase(RegSpec, RegSpec),
+    /// a dereference of the address from summing two registers with offset. for example: `[edi + ecx + 0x40]`
+    RegIndexBaseDisp(RegSpec, RegSpec, i32),
+    /// a dereference of the address held in some register scaled by 1, 2, 4, or 8 with offset. this is almost always used with the `lea` instruction. for example: `[eax * 4 + 0x30]`.
+    RegScaleDisp(RegSpec, u8, i32),
+    /// a dereference of the address from summing a register and index register scaled by 1, 2, 4,
+    /// or 8. for
+    /// example: `[esi + ecx * 4]`
+    RegIndexBaseScale(RegSpec, RegSpec, u8),
+    /// a dereference of the address from summing a register and index register scaled by 1, 2, 4,
+    /// or 8, with offset. for
+    /// example: `[esi + ecx * 4 + 0x1234]`
+    RegIndexBaseScaleDisp(RegSpec, RegSpec, u8, i32),
+    /// an `avx512` dereference of register with optional masking. for example: `[edx]{k3}`
+    RegDerefMasked(RegSpec, RegSpec),
+    /// an `avx512` dereference of register plus offset, with optional masking. for example: `[esp + 0x40]{k3}`
+    RegDispMasked(RegSpec, i32, RegSpec),
+    /// an `avx512` dereference of a register scaled by 1, 2, 4, or 8, with optional masking. this
+    /// seems extraordinarily unlikely to occur in practice. for example: `[esi * 4]{k2}`
+    RegScaleMasked(RegSpec, u8, RegSpec),
+    /// an `avx512` dereference of a register plus index scaled by 1, 2, 4, or 8, with optional masking.
+    /// for example: `[esi + eax * 4]{k6}`
+    RegIndexBaseMasked(RegSpec, RegSpec, RegSpec),
+    /// an `avx512` dereference of a register plus offset, with optional masking.  for example:
+    /// `[esi + eax + 0x1313]{k6}`
+    RegIndexBaseDispMasked(RegSpec, RegSpec, i32, RegSpec),
+    /// an `avx512` dereference of a register scaled by 1, 2, 4, or 8 plus offset, with optional
+    /// masking. this seems extraordinarily unlikely to occur in practice. for example: `[esi *
+    /// 4 + 0x1357]{k2}`
+    RegScaleDispMasked(RegSpec, u8, i32, RegSpec),
+    /// an `avx512` dereference of a register plus index scaled by 1, 2, 4, or 8, with optional
+    /// masking.  for example: `[esi + eax * 4]{k6}`
+    RegIndexBaseScaleMasked(RegSpec, RegSpec, u8, RegSpec),
+    /// an `avx512` dereference of a register plus index scaled by 1, 2, 4, or 8 and offset, with
+    /// optional masking.  for example: `[esi + eax * 4 + 0x1313]{k6}`
+    RegIndexBaseScaleDispMasked(RegSpec, RegSpec, u8, i32, RegSpec),
+    /// no operand. it is a bug for `yaxpeax-x86` to construct an `Operand` of this kind for public
+    /// use; the instruction's `operand_count` should be reduced so as to make this invisible to
+    /// library clients.
+    Nothing,
+}
+
+impl Operand {
+    /// returns `true` if this operand implies a memory access, `false` otherwise.
+    ///
+    /// notably, the `lea` instruction uses a memory operand without actually ever accessing
+    /// memory.
+    pub fn is_memory(&self) -> bool {
+        match self {
+            Operand::DisplacementU16(_)
+            | Operand::DisplacementU32(_)
+            | Operand::DisplacementU64(_)
+            | Operand::RegDeref(_)
+            | Operand::RegDisp(_, _)
+            | Operand::RegScale(_, _)
+            | Operand::RegIndexBase(_, _)
+            | Operand::RegIndexBaseDisp(_, _, _)
+            | Operand::RegScaleDisp(_, _, _)
+            | Operand::RegIndexBaseScale(_, _, _)
+            | Operand::RegIndexBaseScaleDisp(_, _, _, _)
+            | Operand::RegDerefMasked(_, _)
+            | Operand::RegDispMasked(_, _, _)
+            | Operand::RegScaleMasked(_, _, _)
+            | Operand::RegIndexBaseMasked(_, _, _)
+            | Operand::RegIndexBaseDispMasked(_, _, _, _)
+            | Operand::RegScaleDispMasked(_, _, _, _)
+            | Operand::RegIndexBaseScaleMasked(_, _, _, _)
+            | Operand::RegIndexBaseScaleDispMasked(_, _, _, _, _) => true,
+            Operand::ImmediateI8(_)
+            | Operand::ImmediateU8(_)
+            | Operand::ImmediateI16(_)
+            | Operand::ImmediateU16(_)
+            | Operand::ImmediateU32(_)
+            | Operand::ImmediateI32(_)
+            | Operand::ImmediateU64(_)
+            | Operand::ImmediateI64(_)
+            | Operand::Register(_)
+            | Operand::RegisterMaskMerge(_, _, _)
+            | Operand::RegisterMaskMergeSae(_, _, _, _)
+            | Operand::RegisterMaskMergeSaeNoround(_, _, _)
+            | Operand::Nothing => false,
+        }
+    }
+
+    /// return the width of this operand, in bytes. register widths are determined by the
+    /// register's class. the widths of memory operands are recorded on the instruction this
+    /// `Operand` came from; `None` here means the authoritative width is `instr.mem_size()`.
+    pub fn width(&self) -> Option<u8> {
+        match self {
+            Operand::Register(reg) => Some(reg.width()),
+            Operand::RegisterMaskMerge(reg, _, _) => Some(reg.width()),
+            Operand::ImmediateI8(_) | Operand::ImmediateU8(_) => Some(1),
+            Operand::ImmediateI16(_) | Operand::ImmediateU16(_) => Some(2),
+            Operand::ImmediateI32(_) | Operand::ImmediateU32(_) => Some(4),
+            Operand::ImmediateI64(_) | Operand::ImmediateU64(_) => Some(8),
+            // memory operands or `Nothing`
+            _ => None,
+        }
     }
 }
 
@@ -1726,54 +2471,22 @@ impl Opcode {
     /// conditional instructions are `Jcc`, `CMOVcc`, andd `SETcc`.
     pub fn condition(&self) -> Option<ConditionCode> {
         match self {
-            Opcode::JO |
-            Opcode::CMOVO |
-            Opcode::SETO => { Some(ConditionCode::O) },
-            Opcode::JNO |
-            Opcode::CMOVNO |
-            Opcode::SETNO => { Some(ConditionCode::NO) },
-            Opcode::JB |
-            Opcode::CMOVB |
-            Opcode::SETB => { Some(ConditionCode::B) },
-            Opcode::JNB |
-            Opcode::CMOVNB |
-            Opcode::SETAE => { Some(ConditionCode::AE) },
-            Opcode::JZ |
-            Opcode::CMOVZ |
-            Opcode::SETZ => { Some(ConditionCode::Z) },
-            Opcode::JNZ |
-            Opcode::CMOVNZ |
-            Opcode::SETNZ => { Some(ConditionCode::NZ) },
-            Opcode::JA |
-            Opcode::CMOVA |
-            Opcode::SETA => { Some(ConditionCode::A) },
-            Opcode::JNA |
-            Opcode::CMOVNA |
-            Opcode::SETBE => { Some(ConditionCode::BE) },
-            Opcode::JS |
-            Opcode::CMOVS |
-            Opcode::SETS => { Some(ConditionCode::S) },
-            Opcode::JNS |
-            Opcode::CMOVNS |
-            Opcode::SETNS => { Some(ConditionCode::NS) },
-            Opcode::JP |
-            Opcode::CMOVP |
-            Opcode::SETP => { Some(ConditionCode::P) },
-            Opcode::JNP |
-            Opcode::CMOVNP |
-            Opcode::SETNP => { Some(ConditionCode::NP) },
-            Opcode::JL |
-            Opcode::CMOVL |
-            Opcode::SETL => { Some(ConditionCode::L) },
-            Opcode::JGE |
-            Opcode::CMOVGE |
-            Opcode::SETGE => { Some(ConditionCode::GE) },
-            Opcode::JG |
-            Opcode::CMOVG |
-            Opcode::SETG => { Some(ConditionCode::G) },
-            Opcode::JLE |
-            Opcode::CMOVLE |
-            Opcode::SETLE => { Some(ConditionCode::LE) },
+            Opcode::JO | Opcode::CMOVO | Opcode::SETO => Some(ConditionCode::O),
+            Opcode::JNO | Opcode::CMOVNO | Opcode::SETNO => Some(ConditionCode::NO),
+            Opcode::JB | Opcode::CMOVB | Opcode::SETB => Some(ConditionCode::B),
+            Opcode::JNB | Opcode::CMOVNB | Opcode::SETAE => Some(ConditionCode::AE),
+            Opcode::JZ | Opcode::CMOVZ | Opcode::SETZ => Some(ConditionCode::Z),
+            Opcode::JNZ | Opcode::CMOVNZ | Opcode::SETNZ => Some(ConditionCode::NZ),
+            Opcode::JA | Opcode::CMOVA | Opcode::SETA => Some(ConditionCode::A),
+            Opcode::JNA | Opcode::CMOVNA | Opcode::SETBE => Some(ConditionCode::BE),
+            Opcode::JS | Opcode::CMOVS | Opcode::SETS => Some(ConditionCode::S),
+            Opcode::JNS | Opcode::CMOVNS | Opcode::SETNS => Some(ConditionCode::NS),
+            Opcode::JP | Opcode::CMOVP | Opcode::SETP => Some(ConditionCode::P),
+            Opcode::JNP | Opcode::CMOVNP | Opcode::SETNP => Some(ConditionCode::NP),
+            Opcode::JL | Opcode::CMOVL | Opcode::SETL => Some(ConditionCode::L),
+            Opcode::JGE | Opcode::CMOVGE | Opcode::SETGE => Some(ConditionCode::GE),
+            Opcode::JG | Opcode::CMOVG | Opcode::SETG => Some(ConditionCode::G),
+            Opcode::JLE | Opcode::CMOVLE | Opcode::SETLE => Some(ConditionCode::LE),
             _ => None,
         }
     }
