+++ src/long_mode/display.rs
@@ -3,9 +3,9 @@ use core::fmt;
 use yaxpeax_arch::{Colorize, ShowContextual, NoColors, YaxColors};
 use yaxpeax_arch::display::*;
 
-use crate::safer_unchecked::GetSaferUnchecked as _;
 use crate::MEM_SIZE_STRINGS;
-use crate::long_mode::{RegSpec, Opcode, Operand, MergeMode, InstDecoder, Instruction, Segment, PrefixRex, OperandSpec};
+use crate::long_mode::{Opcode, Operand, MergeMode, InstDecoder, Instruction, PrefixRex, OperandSpec};
+use crate::display::regspec_label;
 
 impl fmt::Display for InstDecoder {
     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
@@ -89,54 +89,6 @@ impl fmt::Display for PrefixRex {
     }
 }
 
-impl fmt::Display for Segment {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        match self {
-            Segment::CS => write!(f, "cs"),
-            Segment::DS => write!(f, "ds"),
-            Segment::ES => write!(f, "es"),
-            Segment::FS => write!(f, "fs"),
-            Segment::GS => write!(f, "gs"),
-            Segment::SS => write!(f, "ss"),
-        }
-    }
-}
-
-// register names are grouped by indices scaled by 16.
-// xmm, ymm, zmm all get two indices.
-const REG_NAMES: &[&'static str] = &[
-    "rax", "rcx", "rdx", "rbx", "rsp", "rbp", "rsi", "rdi", "r8", "r9", "r10", "r11", "r12", "r13", "r14", "r15",
-    "eax", "ecx", "edx", "ebx", "esp", "ebp", "esi", "edi", "r8d", "r9d", "r10d", "r11d", "r12d", "r13d", "r14d", "r15d",
-    "ax", "cx", "dx", "bx", "sp", "bp", "si", "di", "r8w", "r9w", "r10w", "r11w", "r12w", "r13w", "r14w", "r15w",
-    "al", "cl", "dl", "bl", "ah", "ch", "dh", "bh", "", "", "", "", "", "", "", "",
-    "al", "cl", "dl", "bl", "spl", "bpl", "sil", "dil", "r8b", "r9b", "r10b", "r11b", "r12b", "r13b", "r14b", "r15b",
-    "cr0", "cr1", "cr2", "cr3", "cr4", "cr5", "cr6", "cr7", "cr8", "cr9", "cr10", "cr11", "cr12", "cr13", "cr14", "cr15",
-    "dr0", "dr1", "dr2", "dr3", "dr4", "dr5", "dr6", "dr7", "dr8", "dr9", "dr10", "dr11", "dr12", "dr13", "dr14", "dr15",
-    "es", "cs", "ss", "ds", "fs", "gs", "", "",
-    "xmm0", "xmm1", "xmm2", "xmm3", "xmm4", "xmm5", "xmm6", "xmm7", "xmm8", "xmm9", "xmm10", "xmm11", "xmm12", "xmm13", "xmm14", "xmm15",
-    "xmm16", "xmm17", "xmm18", "xmm19", "xmm20", "xmm21", "xmm22", "xmm23", "xmm24", "xmm25", "xmm26", "xmm27", "xmm28", "xmm29", "xmm30", "xmm31",
-    "ymm0", "ymm1", "ymm2", "ymm3", "ymm4", "ymm5", "ymm6", "ymm7", "ymm8", "ymm9", "ymm10", "ymm11", "ymm12", "ymm13", "ymm14", "ymm15",
-    "ymm16", "ymm17", "ymm18", "ymm19", "ymm20", "ymm21", "ymm22", "ymm23", "ymm24", "ymm25", "ymm26", "ymm27", "ymm28", "ymm29", "ymm30", "ymm31",
-    "zmm0", "zmm1", "zmm2", "zmm3", "zmm4", "zmm5", "zmm6", "zmm7", "zmm8", "zmm9", "zmm10", "zmm11", "zmm12", "zmm13", "zmm14", "zmm15", "zmm16", "zmm17", "zmm18", "zmm19", "zmm20", "zmm21", "zmm22", "zmm23", "zmm24", "zmm25", "zmm26", "zmm27", "zmm28", "zmm29", "zmm30", "zmm31",
-    "st(0)", "st(1)", "st(2)", "st(3)", "st(4)", "st(5)", "st(6)", "st(7)",
-    "mm0", "mm1", "mm2", "mm3", "mm4", "mm5", "mm6", "mm7",
-    "k0", "k1", "k2", "k3", "k4", "k5", "k6", "k7",
-    "eip", "BUG", "BUG", "BUG", "BUG", "BUG", "BUG", "BUG",
-    "rip", "BUG", "BUG", "BUG", "BUG", "BUG", "BUG", "BUG",
-    "eflags", "BUG", "BUG", "BUG", "BUG", "BUG", "BUG", "BUG",
-    "rflags", "BUG", "BUG", "BUG", "BUG", "BUG", "BUG", "BUG",
-];
-
-pub(crate) fn regspec_label(spec: &RegSpec) -> &'static str {
-    unsafe { REG_NAMES.get_kinda_unchecked((spec.num as u16 + ((spec.bank as u16) << 3)) as usize) }
-}
-
-impl fmt::Display for RegSpec {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        f.write_str(regspec_label(self))
-    }
-}
-
 impl fmt::Display for Operand {
     fn fmt(&self, fmt: &mut fmt::Formatter) -> fmt::Result {
         self.colorize(&NoColors, fmt)
@@ -214,6 +166,9 @@ impl <T: fmt::Write, Y: YaxColors> Colorize<T, Y> for Operand {
                 }
                 f.write_str("{sae}")?;
                 Ok(())
+            }
+            &Operand::DisplacementU16(imm) => {
+                write!(f, "[{}]", colors.address(u16_hex(imm)))
             }
             &Operand::DisplacementU32(imm) => {
                 write!(f, "[{}]", colors.address(u32_hex(imm)))
@@ -446,7 +401,7 @@ fn contextualize_intel<T: fmt::Write, Y: YaxColors>(instr: &Instruction, colors:
     if instr.operand_count > 0 {
         out.write_str(" ")?;
 
-        let x = Operand::from_spec(instr, instr.operands[0]);
+        let x = instr.operand(0);
 
         const RELATIVE_BRANCHES: [Opcode; 21] = [
             Opcode::JMP, Opcode::JRCXZ,
@@ -502,7 +457,7 @@ fn contextualize_intel<T: fmt::Write, Y: YaxColors>(instr: &Instruction, colors:
                         },
                         _ => {
                             out.write_str(", ")?;
-                            let x = Operand::from_spec(instr, instr.operands[i as usize]);
+                            let x = instr.operand(i);
                             if x.is_memory() {
                                 out.write_str(MEM_SIZE_STRINGS[instr.mem_size as usize - 1])?;
                                 out.write_str(" ")?;
@@ -550,7 +505,7 @@ fn contextualize_intel<T: fmt::Write, Y: YaxColors>(instr: &Instruction, colors:
                                     } else {
                                         // this should never be `None` - that would imply two
                                         // memory operands for a broadcasted operation.
-                                        if let Some(width) = Operand::from_spec(instr, instr.operands[i as usize - 1]).width() {
+                                        if let Some(width) = instr.operand(i - 1).width() {
                                             width / instr.mem_size
                                         } else {
                                             0
@@ -952,7 +907,7 @@ impl <T: fmt::Write, Y: YaxColors> ShowContextual<u64, [Option<alloc::string::St
                         }
                     }
                 }
-                let x = Operand::from_spec(self, self.operands[0]);
+                let x = self.operand(0);
                 x.colorize(colors, out)?;
             }
         };
@@ -970,7 +925,7 @@ impl <T: fmt::Write, Y: YaxColors> ShowContextual<u64, [Option<alloc::string::St
                             if let Some(prefix) = self.segment_override_for_op(1) {
                                 write!(out, "{}:", prefix)?;
                             }
-                            let x = Operand::from_spec(self, self.operands[i]);
+                            let x = self.operand(i as u8);
                             x.colorize(colors, out)?
                         }
                     }
